Get UI elements in place
Added names and the ability to change them. A title and
space for a general description or instructions. Radios
for selecting your mark, and a simple play button.
most everything is where i'd like it without the obvious
tweaks. Next step is to get the JS linked to the UI.